+++ R/implied_probabilities.R
@@ -2,23 +2,23 @@
 
 #' Implied probabilities from bookmaker odds.
 #'
-#' This function calculate the implied probabilties from bookmaker odds in decimal format, while
-#' accounting for overround in the odds.
+#' This function calculate the implied probabilities from bookmaker odds in decimal format, while
+#' accounting for over-round in the odds.
 #'
 #' The method 'basic' is the simplest method, and computes the implied probabilities by
 #' dividing the inverted odds by the sum of the inverted odds.
 #'
 #' The methods 'wpo' (Weights Proportional to the Odds), 'or' (Odds Ratio) and 'power' are form the Wisdom of the Crowds document (the updated version) by
-#' Joseph Buchdahl. The method 'or' is origianlly by Cheung (2015), and the method 'power' is there referred
+#' Joseph Buchdahl. The method 'or' is originally by Cheung (2015), and the method 'power' is there referred
 #' to as the logarithmic method.
 #'
 #' The method 'shin' uses the method by Shin (1992, 1993). This model assumes that there is a fraction of
 #' insider trading, and that the bookmakers tries to maximize their profits. In addition to providing
-#' implied probabilties, the method also gives an estimate of the proportion if inside trade, denoted z. Two algorithms
-#' are implemented for finding the probabilities and z. Which algorithm to use is chosen via the shin_mehod argument.
+#' implied probabilities, the method also gives an estimate of the proportion if inside trade, denoted z. Two algorithms
+#' are implemented for finding the probabilities and z. Which algorithm to use is chosen via the shin_method argument.
 #' The default method (shin_method = 'js') is based on the algorithm in Jullien & Salanié (1994). The 'uniroot'
 #' method uses R's built in equation solver to find the probabilities. The uniroot approach is also used for the
-#' 'pwr' and 'or' methods. The two methods might give slightly different answers, especially when the bookamer margin
+#' 'pwr' and 'or' methods. The two methods might give slightly different answers, especially when the bookmaker margin
 #' (and z) is small.
 #'
 #' The 'bb' (short for "balanced books") method is from Fingleton & Waldron (1999), and is a variant of Shin's method. It too assume
@@ -27,22 +27,30 @@
 #'
 #' Both the 'shin' and 'bb' methods can be used together with the 'grossmargin' argument. This is also
 #' from the Fingleton & Waldron (1999) paper, and adds some further assumption to the calculations,
-#' related to opperating costs. grossmargin should be 0 (default) or greater, typical range is 0 to 0.05.
+#' related to operating costs. grossmargin should be 0 (default) or greater, typical range is 0 to 0.05.
 #' For values other than 0, this might sometimes cause some probabilities to not be identifiable. A warning
 #' will be given if this happens.
 #'
 #' The method 'jsd' was developed by Christopher D. Long, and described in a series of Twitter postings
 #' and a python implementation posted on GitHub.
 #'
+#' Methods 'shin', 'or', 'power', and 'jsd' use the uniroot solver to find the correct probabilities. Sometimes it will fail
+#' to find a solution, but it can be made to work by tuning some setting. The uniroot_options argument accepts a list with
+#' options that are passed on to the uniroot function. Currently the interval, maxit, tol and extendInt argument of
+#' uniroot can be changed. See the Troubleshooting vignette for more details.
+#'
 #'
 #' @param odds A matrix or numeric of bookmaker odds. The odds must be in the decimal format.
 #' @param method A string giving the method to use. Valid methods are 'basic', 'shin', 'bb',
 #' 'wpo', 'or', 'power', 'additive', and 'jsd'.
 #' @param normalize Logical. Some of the methods will give small rounding errors. If TRUE (default)
-#' a final normalization is applied to make absoultely sure the
+#' a final normalization is applied to make absolutely sure the
 #' probabilities sum to 1.
+#' @param target_probability Numeric. The value the probabilities should sum to. Default is 1.
 #' @param grossmargin Numeric. Must be 0 or greater. See the details.
-#' @param shin_method Character. Either 'js' (defeault) or 'uniroot'. See the details.
+#' @param shin_method Character. Either 'js' (default) or 'uniroot'. See the details.
+#' @param shin_maxiter numeric. Max number of iterations for shin method 'js'.
+#' @param uniroot_options list. Option passed on to the uniroot solver, for those methods where it is applicable. See 'details'.
 #'
 #'
 #' @return A named list. The first component is named 'probabilities' and contain a matrix of
@@ -52,14 +60,14 @@
 #'  \item{ zvalues (method = 'shin' and method='bb'): The estimated amount of insider trade.}
 #'  \item{ specific_margins (method = 'wpo'): Matrix of the margins applied to each outcome.}
 #'  \item{ odds_ratios (method = 'or'): Numeric with the odds ratio that are used to convert true
-#'  probabilities to bookmaker probabilties.}
+#'  probabilities to bookmaker probabilities.}
 #'  \item{ exponents (method = 'power'): The (inverse) exponents that are used to convert true
-#'  probabilities to bookmaker probabilties.}
+#'  probabilities to bookmaker probabilities.}
 #'  \item{ distance (method = 'jsd'): The Jensen-Shannon distances that are used to convert true
-#'  probabilities to bookmaker probabilties.}
+#'  probabilities to bookmaker probabilities.}
 #' }
 #'
-#' The fourth compnent 'problematic' is a logical vector called indicating if any probabilites has fallen
+#' The fourth component 'problematic' is a logical vector called indicating if any probabilities has fallen
 #' outside the 0-1 range, or if there were some other problem computing the probabilities.
 #'
 #'
@@ -85,20 +93,34 @@
 #'converted_odds$probabilities
 #'
 #' @export
-implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmargin = 0,
-                                  shin_method = 'js'){
+implied_probabilities <- function(odds, method='basic', normalize=TRUE, target_probability = 1,
+                                  grossmargin = 0, shin_method = 'js', shin_maxiter = 1000,
+                                  uniroot_options = NULL){
 
   stopifnot(length(method) == 1,
             tolower(method) %in% c('basic', 'shin', 'bb', 'wpo', 'or', 'power', 'additive', 'jsd'),
             all(odds >= 1, na.rm=TRUE),
+            length(target_probability) == 1,
+            target_probability > 0,
             grossmargin >= 0,
             shin_method %in% c('js', 'uniroot'),
-            length(shin_method) == 1)
+            length(shin_method) == 1,
+            length(shin_maxiter) == 1,
+            shin_maxiter > 1,
+            is.null(uniroot_options) | is.list(uniroot_options))
+
 
   if (method == 'shin' & shin_method == 'uniroot' & grossmargin != 0){
     shin_method <- 'js'
     message('shin_method uniroot does not work when grossmargin is not 0. Method js will be used.')
   }
+
+  if (method == 'shin' & shin_method == 'js' & target_probability != 1){
+    shin_method <- 'uniroot'
+    grossmargin <- 0
+    message('shin_method js does not work when target_probability is not 1. Method uniroot will be used with grossmargin = 0.')
+  }
+
 
   if (!is.matrix(odds)){
 
@@ -109,6 +131,16 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
                      dimnames = list(NULL, names(odds)))
     }
   }
+
+  if (method %in% c('shin', 'or', 'power', 'jsd')){
+    uniroot_opts <- default_uniroot_opts(method = method)
+
+    if (is.list(uniroot_options)){
+      uniroot_opts <- utils::modifyList(uniroot_opts, uniroot_options)
+    }
+
+  }
+
 
   # Prepare the list that will be returned.
   out <- vector(mode='list', length=2)
@@ -121,7 +153,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
   # Inverted odds and margins
   inverted_odds <- 1 / odds
   inverted_odds_sum <- rowSums(inverted_odds)
-  out$margin <- inverted_odds_sum - 1
+  out$margin <- inverted_odds_sum - target_probability
 
   # Missing values
   missing_idx <- apply(odds, MARGIN = 1,
@@ -132,9 +164,15 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
     stop('Some inverse odds sum to less than 1.')
   }
 
+  # Vector to keep track of uniroot problems.
+  use_uniroot <- method %in% c('or', 'power', 'jsd') | (method %in% 'shin' & shin_method == 'uniroot')
+  if (use_uniroot){
+    problematic_uniroot <- logical(n_odds)
+    problematic_uniroot_messages <- character(n_odds)
+  }
 
   if (method == 'basic'){
-    out$probabilities <- inverted_odds / inverted_odds_sum
+    out$probabilities <- (target_probability * inverted_odds) / inverted_odds_sum
 
   } else if (method == 'shin'){
 
@@ -144,7 +182,6 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
     problematic_shin <- logical(n_odds)
 
     if (shin_method == 'js'){
-    #if (shin_method == 'js' | grossmargin != 0){
       for (ii in 1:n_odds){
 
         # Skip rows with missing values.
@@ -155,7 +192,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
         # initialize zz at 0
         zz_tmp <- 0
 
-        for (jj in 1:1000){
+        for (jj in 1:shin_maxiter){
           zz_prev <- zz_tmp
 
           if (grossmargin != 0){
@@ -166,7 +203,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
 
           if (abs(zz_tmp - zz_prev)  <= .Machine$double.eps^0.25){
             break
-          } else if (jj >= 1000){
+          } else if (jj >= shin_maxiter){
             problematic_shin[ii] <- TRUE
           }
 
@@ -174,7 +211,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
           probs[ii,] <- shin_func(zz=zz_tmp, io = inverted_odds[ii,])
         }
       }
-    } else {
+    } else if (shin_method == 'uniroot'){
       for (ii in 1:n_odds){
 
         # Skip rows with missing values.
@@ -182,10 +219,18 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
           next
         }
 
-        res <- stats::uniroot(f=shin_solvefor, interval = c(0,0.4), io=inverted_odds[ii,])
+
+        res <- uniroot2(f=shin_solvefor, io=inverted_odds[ii,], trgtprob = target_probability,
+                        interval = uniroot_opts$interval, extendInt = uniroot_opts$extendInt,
+                        tol = uniroot_opts$tol, maxiter = uniroot_opts$maxiter)
 
         zvalues[ii] <- res$root
         probs[ii,] <- shin_func(zz=res$root, io = inverted_odds[ii,])
+
+        if (!is.null(res$message)){
+          problematic_uniroot[ii] <- TRUE
+          problematic_uniroot_messages[ii] <- res$message
+        }
 
       }
     }
@@ -194,13 +239,14 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
     out$zvalues <- zvalues
 
     if (any(problematic_shin[!missing_idx])){
-      warning(sprintf('Could not find z: Did not converge in %d instances. Some results may be unreliable. See the "problematic" vector in the output.',
+      warning(sprintf('Could not find z: Did not converge in %d instances. Some results may be unreliable.',
                       sum(problematic_shin)))
     }
 
   } else if (method == 'bb'){
 
-    zz <- (((1-grossmargin)*inverted_odds_sum) - 1) / (n_outcomes-1)
+    # zz <- (((1-grossmargin) * inverted_odds_sum) - 1) / (n_outcomes-1)
+    zz <- (((1-grossmargin) * inverted_odds_sum) - target_probability) / (n_outcomes-target_probability)
     probs <- (((1-grossmargin) * inverted_odds) - zz) / (1-zz)
 
     out$probabilities <- probs
@@ -225,9 +271,18 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
         next
       }
 
-      res <- stats::uniroot(f=or_solvefor, interval = c(0.05, 5), io=inverted_odds[ii,])
+      res <- uniroot2(f=or_solvefor, io=inverted_odds[ii,], trgtprob = target_probability,
+                      interval = uniroot_opts$interval, extendInt = uniroot_opts$extendInt,
+                      tol = uniroot_opts$tol, maxiter = uniroot_opts$maxiter)
+
       odds_ratios[ii] <- res$root
       probs[ii,] <- or_func(cc=res$root, io = inverted_odds[ii,])
+
+      if (!is.null(res$message)){
+        problematic_uniroot[ii] <- TRUE
+        problematic_uniroot_messages[ii] <- res$message
+      }
+
     }
 
     out$probabilities <- probs
@@ -245,9 +300,16 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
         next
       }
 
-      res <- stats::uniroot(f=pwr_solvefor, interval = c(0.0001, 1), io=inverted_odds[ii,])
+      res <- uniroot2(f=pwr_solvefor, io=inverted_odds[ii,], trgtprob = target_probability,
+                      interval = uniroot_opts$interval, extendInt = uniroot_opts$extendInt,
+                       tol = uniroot_opts$tol, maxiter = uniroot_opts$maxiter)
       exponents[ii] <- res$root
       probs[ii,] <- pwr_func(nn=res$root, io = inverted_odds[ii,])
+
+      if (!is.null(res$message)){
+        problematic_uniroot[ii] <- TRUE
+        problematic_uniroot_messages[ii] <- res$message
+      }
     }
 
     out$probabilities <- probs
@@ -264,7 +326,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
         next
       }
 
-      probs[ii,] <- inverted_odds[ii,] - ((inverted_odds_sum[ii] - 1) / n_outcomes)
+      probs[ii,] <- inverted_odds[ii,] - ((inverted_odds_sum[ii] - target_probability) / n_outcomes)
     }
 
     out$probabilities <- probs
@@ -280,25 +342,28 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
         next
       }
 
-      # 0.1 seems to be a reasonable upper bound, with possibility of extending.
-      res <- stats::uniroot(f=jsd_solvefor, interval = c(0.0000001, 0.1),
-                            io=inverted_odds[ii,],
-                            tol=0.0000001)
 
+      res <- uniroot2(f=jsd_solvefor, io=inverted_odds[ii,], trgtprob = target_probability,
+                      interval = uniroot_opts$interval, extendInt = uniroot_opts$extendInt,
+                      tol = uniroot_opts$tol, maxiter = uniroot_opts$maxiter)
       jsds[ii] <- res$root
       probs[ii,] <- jsd_func(jsd=res$root, io = inverted_odds[ii,])
+
+      if (!is.null(res$message)){
+        problematic_uniroot[ii] <- TRUE
+        problematic_uniroot_messages[ii] <- res$message
+      }
     }
 
     out$probabilities <- probs
     out$distance <- jsds
 
-
   }
 
-  ## do a final normalization to make sure the probabilites
+  ## do a final normalization to make sure the probabilities
   ## sum to 1 without rounding errors.
   if (normalize){
-    out$probabilities <- out$probabilities / rowSums(out$probabilities)
+    out$probabilities <- (target_probability * out$probabilities) / rowSums(out$probabilities)
   }
 
   # Make sure the matrix of implied probabilities has column names.
@@ -306,7 +371,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
     colnames(out$probabilities) <- colnames(odds)
   }
 
-  # check if there are any probabilites outside the 0-1 range.
+  # check if there are any probabilities outside the 0-1 range.
   problematic <- apply(out$probabilities, MARGIN = 1, FUN=function(x){any(x > 1 | x < 0)})
   problematic[is.na(problematic)] <- TRUE
   problematic[missing_idx] <- NA
@@ -316,6 +381,20 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
                     sum(problematic)))
   }
 
+  # Give warnings for problems when uniroot was used.
+  if (use_uniroot){
+
+    if (any(problematic_uniroot)){
+      problematic[problematic_uniroot] <- TRUE
+
+      problematic_uniroot_messages <- unique(problematic_uniroot_messages)
+      for (ww in 1:length(problematic_uniroot_messages)){
+        warning(problematic_uniroot_messages[ww])
+      }
+    }
+  }
+
+
   if (method == 'shin'){
     problematic <- problematic | problematic_shin
   }
@@ -323,7 +402,7 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
   if (method %in% c('shin', 'bb')){
     negative_z <- out$zvalues < 0
     if (any(negative_z[!missing_idx])){
-      warning(sprintf('z estimated to be negative: Some results may be unreliable. See the "problematic" vector in the output.',
+      warning(sprintf('z estimated to be negative: Some results may be unreliable.',
                       negative_z))
     }
   }
@@ -339,6 +418,48 @@ implied_probabilities <- function(odds, method='basic', normalize=TRUE, grossmar
 # Internal functions used to transform probabilities
 # and be used with uniroot.
 #########################################################
+
+
+default_uniroot_opts <- function(method){
+  opts <- list(extendInt = 'yes',
+               maxiter = 1000,
+               tol = .Machine$double.eps^0.25)
+
+  if (method == 'shin'){
+    opts$interval <- c(0, 0.4)
+  } else if (method == 'or'){
+    opts$interval <- c(0.95, 5)
+  } else if (method == 'power'){
+    opts$interval <- c(0.0001, 1)
+  } else if (method == 'jsd'){
+    opts$interval <- c(0.0000001, 0.1)
+    opts$tol = 0.000001
+  }
+
+  return(opts)
+
+}
+
+
+
+# Wrapper around uniroot, but returns a list with NA results
+# if the solver fails.
+uniroot2 <- function(f, interval, ...,
+                     extendInt = 'no', tol = .Machine$double.eps^0.25, maxiter = 1000){
+
+
+  res <- tryCatch({
+    stats::uniroot(f = f, ..., interval = interval, extendInt = extendInt, tol = tol, maxiter = maxiter)
+  }, error = function(e){
+    list(root = NA, message = as.character(e))
+  })
+
+  return(res)
+
+}
+
+
+
 
 # Calculate the probabilities using Shin's formula, for a given value of z.
 # io = inverted odds.
@@ -358,17 +479,17 @@ or_func <- function(cc, io){
 
 # the condition that the sum of the probabilites must sum to 1.
 # Used with uniroot.
-shin_solvefor <- function(zz, io){
+shin_solvefor <- function(zz, io, trgtprob){
   tmp <- shin_func(zz, io)
-  1 - sum(tmp) # 0 when the condition is satisfied.
+  sum(tmp) - trgtprob # 0 when the condition is satisfied.
 }
 
 # The condition that the sum of the probabilites must sum to 1.
 # This function calulates the true probability, given bookmaker
 # probabilites xx, and the odds ratio cc.
-or_solvefor <- function(cc, io){
+or_solvefor <- function(cc, io, trgtprob){
   tmp <- or_func(cc, io)
-  sum(tmp) - 1
+  sum(tmp) - trgtprob
 }
 
 # power function.
@@ -379,9 +500,9 @@ pwr_func <- function(nn, io){
 # The condition that the sum of the probabilites must sum to 1.
 # This function calulates the true probability, given bookmaker
 # probabilites xx, and the inverse exponent. nn.
-pwr_solvefor <- function(nn, io){
+pwr_solvefor <- function(nn, io, trgtprob){
   tmp <- pwr_func(nn, io)
-  sum(tmp) - 1
+  sum(tmp) - trgtprob
 }
 
 # Simple discrete KL-divergence.
@@ -415,8 +536,10 @@ jsd_func <- function(jsd, io){
     # That the underlying probability i less than the
     # inverse odds.
     pp[ii] <- stats::uniroot(f = tosolve,
-                      interval = c(0.00001, io[ii]),
+                      interval = c(0.000001, io[ii]),
+                      extendInt = 'yes',
                       io = io[ii], jsd = jsd)$root
+
   }
   return(pp)
 }
@@ -424,8 +547,8 @@ jsd_func <- function(jsd, io){
 # Calculate the probabilities using the Jensen-Shannon distance method,
 # for a given value of the odds ratio cc.
 # io = inverted odds.
-jsd_solvefor <- function(jsd, io){
-  sum(jsd_func(jsd=jsd, io = io)) - 1
+jsd_solvefor <- function(jsd, io, trgtprob){
+  sum(jsd_func(jsd=jsd, io = io)) - trgtprob
 }