+++ NEWS.md
@@ -1,4 +1,13 @@
 
+
+# implied Version 0.5
+* Can now convert odds to probabilities that should sum to other values than 1, using the target_probability argument in implied_probabilities().
+* New vignette 'Troubleshooting'.
+* Introduction vignette updated with how to use the new target_probability option.
+* New option 'uniroot_options' in implied_probabilities, to better control the uniroot solver.
+* Fixed many spelling errors in the documentation.
+
+
 # implied Version 0.4.1
 * Small change to how the 'jsd' method in implied_probabilities() works, so that it works in some cases where it used to fail.
 * Fixed a link in the Introduction vignette.